refactor: 💡 extract expand components from row file
