+++ src/Molecules/FlexTable/Row/components/index.ts
@@ -1,0 +1,3 @@
+export { ExpandItem, ExpandItems } from './ExpandItems';
+export { ExpandElement } from './ExpandElement';
+export { ExpandArea } from './ExpandArea';
